+++ plugins/wme-sitebuilder/CHANGELOG.md
@@ -5,3 +5,5 @@
 ### Minor Changes
 
 - fcadf1d: Consolidate the WME Sitebuilder WordPress plugin and library into the same repository with other Sitebuilder packages.
+- Updated dependencies [0413284]
+  - @moderntribe/sitebuilder@1.7.0